requirements.txt
fastapi
motor
pydantic
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
uvicorn
langchain-groq